FIx MaxDepth/Regex issues
Fixed the MaxDepth parameter of the PathSearchOnFPGA Tcl command so that it works correctly. Also added a new parameter to match port names exactly (false by default). Finally, added a bit more feedback on the PathSearchOnFPGA command if any of the regular expressions return no matches or if no routes were found.